refactor(nav): extract the drawer out of :shared into :feature:conversations
The navigation drawer had grown into a stealth feature living in :shared —
DrawerContent, AccountSwitcherSheet, ConversationListStateHolder, drawer
string resources, and a ~700-line NavHostViewModel. Move the drawer feature
into feature/conversations/drawer/ so :shared stays nav glue, and split the
drawer-data concerns out of NavHostViewModel into a new activity-scoped
DrawerViewModel. The cross-module compile edge between the drawer and
feature:conversations disappears; :shared sheds ~2,000 lines.

Also fix the search-mode staleness this coupling hid: ConversationListViewModel
dropped every Room emission during an active search, so a conversation deleted
from the drawer stayed visible in the list screen's search results. Active
search now reconciles visible results against Room — dropping rows that left
Room and refreshing rows Room still holds by updatedAt — without pulling
non-matching Room rows into server-sourced results.
