+++ feature/conversations/src/androidUnitTest/kotlin/com/garfiec/librechat/feature/conversations/di/ConversationsModuleVerificationTest.kt
@@ -2,7 +2,9 @@ package com.garfiec.librechat.feature.conversations.di
 
 import android.app.Application
 import android.content.Context
+import com.garfiec.librechat.core.common.identity.ActiveAccountProvider
 import com.garfiec.librechat.core.data.datastore.ServerDataStore
+import com.garfiec.librechat.core.data.datastore.SettingsDataStore
 import com.garfiec.librechat.core.data.repository.ConfigRepository
 import com.garfiec.librechat.core.data.repository.ConversationRepository
 import com.garfiec.librechat.core.data.repository.MessageRepository
@@ -32,6 +34,9 @@ class ConversationsModuleVerificationTest {
                 RoleRepository::class,
                 ServerDataStore::class,
                 CoroutineDispatcher::class,
+                // DrawerViewModel deps sourced from other modules.
+                ActiveAccountProvider::class,
+                SettingsDataStore::class,
             ),
         )
     }
