+++ feature/conversations/src/commonMain/composeResources/values-pt/strings.xml
@@ -76,4 +76,12 @@
 
     <!-- Conversation item fallback -->
     <string name="chat">Conversa</string>
+    <!-- Drawer strings (moved from :shared) -->
+    <string name="agents">Agentes</string>
+    <string name="favorites">Favoritos</string>
+    <string name="files">Arquivos</string>
+    <string name="no_conversations_found">Nenhuma conversa encontrada</string>
+    <string name="search_conversations_placeholder">Buscar conversas…</string>
+    <string name="settings">Configurações</string>
+    <string name="skills">Habilidades</string>
 </resources>
