+++ feature/conversations/src/commonMain/composeResources/values-zh/strings.xml
@@ -76,4 +76,12 @@
 
     <!-- Conversation item fallback -->
     <string name="chat">聊天</string>
+    <!-- Drawer strings (moved from :shared) -->
+    <string name="agents">智能体</string>
+    <string name="favorites">收藏</string>
+    <string name="files">文件</string>
+    <string name="no_conversations_found">未找到对话</string>
+    <string name="search_conversations_placeholder">搜索对话…</string>
+    <string name="settings">设置</string>
+    <string name="skills">技能</string>
 </resources>
